refactor(cms): split content group menu components
Separated ContentGroupsMenuItems logic into distinct components:
GroupMenu, GroupContentModels, HasContentEntryPermissions, and Icon.
Improves code readability and maintainability by modularizing menu item
rendering and permission checks for CMS content groups.

Ref: WEB-5362
