+++ packages/app-headless-cms/src/admin/menus/HasContentEntryPermissions.ts
@@ -1,0 +1,37 @@
+import type { CmsGroup, CmsModel } from "~/types.js";
+import { usePermission } from "~/admin/hooks/usePermission.js";
+
+interface HasContentEntryPermissionsProps {
+    group: CmsGroup;
+    contentModel?: CmsModel;
+    children: JSX.Element;
+}
+
+const HasContentEntryPermissions = ({
+    group,
+    contentModel,
+    children
+}: HasContentEntryPermissionsProps) => {
+    const { canReadEntries } = usePermission();
+
+    if (contentModel) {
+        if (!canReadEntries({ contentModelGroup: group, contentModel })) {
+            return null;
+        }
+    } else {
+        const hasContentEntryPermission = group.contentModels.some(contentModel =>
+            canReadEntries({
+                contentModelGroup: group,
+                contentModel
+            })
+        );
+
+        if (group.contentModels.length > 0 && !hasContentEntryPermission) {
+            return null;
+        }
+    }
+
+    return children;
+};
+
+export default HasContentEntryPermissions;
